+++ src/components/home/index.jsx
@@ -5,7 +5,7 @@ import styles from "components/home/index.module.scss";
 import Button from "components/base/button";
 import CreateTask from "components/task/create-new";
 import TaskCard from "components/task/existing-card";
-import FilterBtnContainer from "components/filter";
+import FilterBtnContainer from "components/filter-btn-container";
 
 function HomeContainer() {
   const tasks = useSelector((state) => state.todo.tasks);
